Fix visual regression tests
